# docs/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nest Project API",
    "version": "1.0.0",
    "description": "OpenAPI document generated from the current NestJS controllers and DTOs. Import this file into Apifox."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local NestJS server"
    }
  ],
  "tags": [
    {
      "name": "App"
    },
    {
      "name": "User"
    },
    {
      "name": "Question"
    },
    {
      "name": "Bank"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "App"
        ],
        "summary": "Health check",
        "operationId": "getHello",
        "security": [],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessString"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all users",
        "operationId": "findAllUsers",
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create user",
        "operationId": "createUser",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              },
              "example": {
                "name": "test",
                "phone": "13800138000",
                "password": "123456",
                "confirmPassword": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/user/search": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Search user by id",
        "operationId": "searchUser",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Login",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "example": {
                "phone": "13800138000",
                "password": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/user/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Update user",
        "operationId": "updateUser",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              },
              "example": {
                "name": "new-name",
                "phone": "13800138001",
                "password": "654321"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/question": {
      "get": {
        "tags": [
          "Question"
        ],
        "summary": "Get questions",
        "operationId": "getQuestions",
        "parameters": [
          {
            "$ref": "#/components/parameters/CollectionTypeQuery"
          },
          {
            "$ref": "#/components/parameters/QuestionTypeQuery"
          },
          {
            "name": "content",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bank",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bankId",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "discipline",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "disciplineId",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "number",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": [
          "Question"
        ],
        "summary": "Create one or many questions",
        "operationId": "createQuestion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CreateQuestionDto"
                  },
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/CreateQuestionDto"
                    }
                  }
                ]
              },
              "examples": {
                "single": {
                  "value": {
                    "type": "SingleChoice",
                    "content": "1 + 1 = ?",
                    "options": [
                      {
                        "key": "A",
                        "text": "1"
                      },
                      {
                        "key": "B",
                        "text": "2"
                      }
                    ],
                    "answer": "B",
                    "bank": "默认题库",
                    "discipline": "数学"
                  }
                },
                "many": {
                  "value": [
                    {
                      "type": "TrueFalse",
                      "content": "太阳从东方升起",
                      "options": [
                        {
                          "key": "A",
                          "text": "正确"
                        },
                        {
                          "key": "B",
                          "text": "错误"
                        }
                      ],
                      "answer": "A",
                      "bank": "默认题库",
                      "discipline": "常识"
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/question/{id}": {
      "get": {
        "tags": [
          "Question"
        ],
        "summary": "Get question detail",
        "operationId": "getQuestion",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/question/collection": {
      "get": {
        "tags": [
          "Question"
        ],
        "summary": "Get collections",
        "operationId": "getCollection",
        "parameters": [
          {
            "$ref": "#/components/parameters/CollectionTypeQuery"
          },
          {
            "name": "bankId",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "disciplineId",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isDay",
            "in": "query",
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1
              ]
            }
          },
          {
            "name": "questionType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/QuestionType"
            }
          },
          {
            "name": "detailed",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": [
          "Question"
        ],
        "summary": "Create collection",
        "operationId": "createCollection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionDto"
              },
              "example": {
                "questionId": 1,
                "type": "Note"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Question"
        ],
        "summary": "Delete collections by filters",
        "operationId": "deleteAllCollections",
        "parameters": [
          {
            "$ref": "#/components/parameters/CollectionTypeQuery"
          },
          {
            "name": "bankId",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "disciplineId",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/question/collection/{id}": {
      "delete": {
        "tags": [
          "Question"
        ],
        "summary": "Delete collection by id",
        "operationId": "deleteCollection",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/question/collection/exist": {
      "get": {
        "tags": [
          "Question"
        ],
        "summary": "Check whether collection exists",
        "operationId": "isCollectionExist",
        "parameters": [
          {
            "name": "questionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/question/check": {
      "post": {
        "tags": [
          "Question"
        ],
        "summary": "Check one or many answers",
        "operationId": "checkAnswer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CheckAnswerDto"
                  },
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/CheckAnswerDto"
                    }
                  }
                ]
              },
              "example": {
                "questionId": 1,
                "answer": "B"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/question/resolution": {
      "get": {
        "tags": [
          "Question"
        ],
        "summary": "Get answer resolutions",
        "operationId": "getResolutions",
        "parameters": [
          {
            "name": "bankId",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "bankName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "disciplineName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "disciplineId",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/bank": {
      "get": {
        "tags": [
          "Bank"
        ],
        "summary": "Get bank list",
        "operationId": "getBankList",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "creator",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/bank/create": {
      "post": {
        "tags": [
          "Bank"
        ],
        "summary": "Create bank",
        "operationId": "createBank",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBankDto"
              },
              "example": {
                "name": "默认题库",
                "description": "示例题库",
                "creator": "test",
                "disciplines": [
                  "数学",
                  "常识"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/bank/{id}": {
      "get": {
        "tags": [
          "Bank"
        ],
        "summary": "Get bank detail",
        "operationId": "getBankDetail",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/bank/discipline": {
      "post": {
        "tags": [
          "Bank"
        ],
        "summary": "Create discipline",
        "operationId": "createDiscipline",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDisciplineDto"
              },
              "examples": {
                "single": {
                  "value": {
                    "discipline": "数学"
                  }
                },
                "many": {
                  "value": {
                    "discipline": [
                      "数学",
                      "英语"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessAny"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int32"
        }
      },
      "QuestionTypeQuery": {
        "name": "type",
        "in": "query",
        "schema": {
          "$ref": "#/components/schemas/QuestionType"
        }
      },
      "CollectionTypeQuery": {
        "name": "type",
        "in": "query",
        "schema": {
          "$ref": "#/components/schemas/CollectionType"
        }
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "nullable": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "example": 401
          },
          "message": {
            "type": "string",
            "example": "Unauthorized"
          },
          "data": {
            "nullable": true,
            "example": null
          }
        }
      },
      "QuestionType": {
        "type": "string",
        "enum": [
          "SingleChoice",
          "MultiChoice",
          "TrueFalse"
        ]
      },
      "Answer": {
        "type": "string",
        "enum": [
          "A",
          "B",
          "C",
          "D"
        ]
      },
      "CollectionType": {
        "type": "string",
        "enum": [
          "Mistake",
          "Note"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "required": [
          "name",
          "phone",
          "password",
          "confirmPassword"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "minLength": 11,
            "maxLength": 11
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 12,
            "format": "password"
          },
          "confirmPassword": {
            "type": "string",
            "minLength": 6,
            "maxLength": 12,
            "format": "password"
          }
        }
      },
      "LoginDto": {
        "type": "object",
        "required": [
          "phone",
          "password"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "minLength": 11,
            "maxLength": 11
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "minLength": 11,
            "maxLength": 11
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 12,
            "format": "password"
          }
        }
      },
      "OptionDto": {
        "type": "object",
        "required": [
          "key",
          "text"
        ],
        "properties": {
          "key": {
            "$ref": "#/components/schemas/Answer"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "CreateQuestionDto": {
        "type": "object",
        "required": [
          "type",
          "content",
          "options",
          "answer",
          "bank",
          "discipline"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "content": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "minItems": 2,
            "items": {
              "$ref": "#/components/schemas/OptionDto"
            }
          },
          "answer": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Answer"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            ]
          },
          "bank": {
            "type": "string"
          },
          "discipline": {
            "type": "string"
          }
        }
      },
      "CheckAnswerDto": {
        "type": "object",
        "required": [
          "questionId",
          "answer"
        ],
        "properties": {
          "questionId": {
            "type": "number"
          },
          "answer": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Answer"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            ]
          }
        }
      },
      "CreateCollectionDto": {
        "type": "object",
        "required": [
          "questionId"
        ],
        "properties": {
          "questionId": {
            "type": "number"
          },
          "type": {
            "$ref": "#/components/schemas/CollectionType"
          }
        }
      },
      "CreateBankDto": {
        "type": "object",
        "required": [
          "name",
          "description",
          "creator",
          "disciplines"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "creator": {
            "type": "string"
          },
          "disciplines": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateDisciplineDto": {
        "type": "object",
        "required": [
          "discipline"
        ],
        "properties": {
          "discipline": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          }
        }
      }
    },
    "responses": {
      "SuccessAny": {
        "description": "Success",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            }
          }
        }
      },
      "SuccessString": {
        "description": "Success",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "Error": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}